# Conflicts:
#	common/msg/glmsg.go
#	gameServer/center/internal/handler.go
#	gameServer/center/internal/module.go
#	gameServer/common/mj/mj_base/baselogic.go
#	gameServer/common/pk/pk_base/pk_data_mgr.go
#	gameServer/common/pk/pk_interface.go
#	gameServer/db/model/create_room_info.go
#	hallServer/center/external.go
#	hallServer/center/internal/handler.go
#	hallServer/game_list/internal/handler.go
#	hallServer/hallApp/hallServer.json

+++ gameServer/center/internal/module.go
@@ -13,8 +13,8 @@ import (
 var (
 	skeleton   = base.NewSkeleton()
 	ChanRPC    = skeleton.ChanRPCServer
-	Users      = make(map[int]*chanrpc.Server) //本服玩家
-	OtherUsers = make(map[int]string)          //其他服登录的玩家  key is uid， values is NodeId
+	Users      = make(map[int64]*chanrpc.Server) //本服玩家
+	OtherUsers = make(map[int64]string)          //其他服登录的玩家  key is uid， values is NodeId
 )
 
 type Module struct {
